Extract SearchBuilder behavior from classes into modules This should help support blacklight generating a search_builder into the local application:
```ruby
class SearchBuilder
  include Blacklight::SearchBuilder
  include Blacklight::Solr::SearchBuilder
end
```
