+++ lib/blacklight.rb
@@ -2,11 +2,12 @@
 require 'kaminari'
 require 'deprecation'
 module Blacklight
-  autoload :Configurable, 'blacklight/configurable'
-  autoload :Configuration, 'blacklight/configuration'
-  autoload :SearchFields, 'blacklight/search_fields'
-  autoload :SearchBuilder, 'blacklight/search_builder'
-  
+  autoload :Configurable,          'blacklight/configurable'
+  autoload :Configuration,         'blacklight/configuration'
+  autoload :SearchFields,          'blacklight/search_fields'
+  autoload :SearchBuilder,         'blacklight/search_builder'
+  autoload :SearchBuilderBehavior, 'blacklight/search_builder_behavior'
+
   autoload :Document, 'blacklight/document'
   autoload :Solr, 'blacklight/solr'
